beta04 added resizing image

+++ mediapicker/src/main/java/iam/thevoid/mediapicker/picker/metrics/Resolution.kt
@@ -1,3 +1,7 @@
 package iam.thevoid.mediapicker.picker.metrics
 
-data class Resolution @JvmOverloads constructor(val width: Long = -1, val height: Long = -1)
+/**
+ * @param width: width in pixels
+ * @param height: height in pixels
+ */
+data class Resolution @JvmOverloads constructor(val width: Int = -1, val height: Int = -1)
